+++ packages/esbuild-config/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@ooz/esbuild-config",
+  "packageManager": "yarn@4.4.0",
+  "main": "index.js",
+  "files": [
+    "index.js"
+  ],
+  "scripts": {
+    "build": "echo \"Tada!\"",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "esbuild": "0.16.17"
+  }
+}
