+++ packages/themes/build.js
@@ -1,40 +1,6 @@
-import esbuild from "esbuild";
-import pkg from "./package.json" assert { type: "json" };
+import run from '@ooz/esbuild-config';
+import pkg from './package.json' assert { type: 'json' };
 
-const dev = process.argv.includes("--dev");
-const minify = !dev;
-
-const watch = process.argv.includes("--watch");
-
-const external = Object.keys({
-  ...pkg.dependencies,
-  ...pkg.peerDependencies,
-});
-
-const baseConfig = {
-  entryPoints: ["src/index.ts"],
-  bundle: true,
-  minify,
-  sourcemap: true,
-  outdir: "dist",
-  target: "es2019",
-  watch,
-  external,
-};
-
-Promise.all([
-  esbuild.build({
-    ...baseConfig,
-    format: "esm",
-  }),
-  esbuild.build({
-    ...baseConfig,
-    format: "cjs",
-    outExtension: {
-      ".js": ".cjs",
-    },
-  }),
-]).catch(() => {
-  console.error("Build failed");
-  process.exit(1); // node 프로세스 종료
-});
+run({
+  pkg,
+})
