+++ packages/themes/package.json
@@ -19,7 +19,7 @@
     "clean": "rm -rf dist"
   },
   "devDependencies": {
-    "esbuild": "^0.16.17",
+    "@ooz/esbuild-config": "workspace:^",
     "typescript": "^5.5.4"
   }
 }
